requirements.txt
vosk
sounddevice
numpy
pyyaml
requests
